requirements.txt
mkdocs==1.6.1
mkdocs-material==9.5.34
mkdocs-glightbox>=0.5
pymdown-extensions==10.9
mkdocs-git-revision-date-localized-plugin==1.2.4
mkdocs-macros-plugin
PyYAML
Jinja2
ruamel.yaml